+++ lib/src/client/http_client_message_handler.dart
@@ -1,0 +1,9 @@
+import 'http_client.dart';
+
+/// A function that is called when a response is received from the HTTP server.
+typedef HttpClientMessageHandler =
+    void Function(HttpClientResponseData response);
+
+/// A function that is called periodically to update the client.
+typedef HttpClientOnTick = void Function(HttpClient client);
+