+++ lib/src/core/service/service.dart
@@ -1,0 +1,7 @@
+/// Represents a service that can be used to process requests and generate responses.
+abstract class Service<RequestType, ResponseType> {
+
+  /// Processes a request and returns a response.
+  ResponseType process(RequestType request);
+
+}